backend/requirements.txt
Flask==2.3.3
psycopg[binary]==3.2.9
flasgger==0.9.7.1
flask-cors==4.0.0
python-dotenv==1.0.0 